% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Function that retruns a plot of the given country's population over the years 1950:2020}
\usage{
CountryPopulation(nameOfCountry)
}
\description{
@format A function that returns a ggplot2 graph of the given country's population
 \describe{
   \item{nameOfCountry}{A string that represents the name of the desired country.}
   \item{return}{A ggplot2 graph that displays the given countrys population data.}
 }
}
