% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World cup average attendance per country and year}
\format{
a data frame with 22 observations of 4 variables
\describe{
   \item{Matches}{How many matches have been played.}
   \item{Totalattendance}{What is the total attendance for this country.}
   \item{Averageattendance}{What is the average attendance for this country.}
   \item{WorldCup}{The country and year the world cup took place.}
}
}
\usage{
World_Cup
}
\description{
World cup average attendance per country and year
}
\keyword{datasets}
