% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World population by country and year}
\format{
a data frame with 235 observations of 72 variables
\describe{
   \item{Region, subregion, country or area}{The country of the observation.}
   \item{1950:2020}{The year population of the country.}
}
}
\usage{
WorldPopulation
}
\description{
World population by country and year
}
\keyword{datasets}
